pyproject.toml
[project]
name = "master-control"
version = "2.0.0"
description = "A robust access control system for hackerspaces, managing RFID fobs, database synchronization, and physical door hardware via AMQP and Serial communication."
authors = [
    { name = "Rohan Kar", email = "rohan@example.com" }
]
dependencies = [
    "sqlalchemy>=2.0.0",
    "pyserial>=3.5",
    "mysql-connector-python>=8.0.33",
    "observable>=1.0.3",
    "pika>=1.3.1",
    "configparser>=5.3.0",
    "python-dotenv>=1.0.0"
]
requires-python = ">=3.9"

[project.scripts]
master-control = "mastercontrol:main"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
strict = true
ignore_missing_imports = true
